move cli test and change pyproject.toml entry point
