+++ tests/autorag/test_evaluator.py
@@ -1,6 +1,5 @@
 import os.path
 import pathlib
-import subprocess
 import tempfile
 
 import pandas as pd
@@ -125,23 +124,6 @@ def test_start_trial(evaluator):
     assert trial_summary_df['best_module_name'][0] == 'bm25'
     assert trial_summary_df['best_module_params'][0] == {'top_k': 10}
     assert trial_summary_df['best_execution_time'][0] > 0
-
-
-def test_evaluator_cli():
-    with tempfile.TemporaryDirectory() as project_dir:
-        result = subprocess.run(['autorag', 'evaluate', '--config', os.path.join(resource_dir, 'simple.yaml'),
-                                 '--qa_data_path', os.path.join(resource_dir, 'qa_data_sample.parquet'),
-                                 '--corpus_data_path', os.path.join(resource_dir, 'corpus_data_sample.parquet'),
-                                 '--project_dir', project_dir])
-        assert result.returncode == 0
-        # check if the files are created
-        assert os.path.exists(os.path.join(project_dir, '0'))
-        assert os.path.exists(os.path.join(project_dir, 'data'))
-        assert os.path.exists(os.path.join(project_dir, 'resources'))
-        assert os.path.exists(os.path.join(project_dir, 'trial.json'))
-        assert os.path.exists(os.path.join(project_dir, '0', 'retrieve_node_line'))
-        assert os.path.exists(os.path.join(project_dir, '0', 'retrieve_node_line', 'retrieval'))
-        assert os.path.exists(os.path.join(project_dir, '0', 'retrieve_node_line', 'retrieval', '0.parquet'))
 
 
 def start_trial_full(evaluator):
